requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
pydantic>=1.8.0
transformers>=4.30.0
torch>=2.0.0
redis>=4.5.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0 